updata tSNE FP
